backend_fastapi/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
beanie>=1.18.0
motor>=3.1.1
httpx>=0.24.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0